Refined.Extras.Predicates.Text

Description

Predicates for Text and String.

Since: 0.1.0.0

Symbol Equality

data SymEqualTo (c :: Symbol)

Predicate equality for symbols.

Examples

Expand

>>> validate @(SymEqualTo "c") Proxy 'c'
Nothing

>>> showRefineException <$> validate @(SymEqualTo "abc") Proxy 'c'
Just "RefineOtherException (SymEqualTo \"abc\") \"c is not a single Char\""

>>> validate @(SymEqualTo "abc") Proxy "abc"
Nothing

>>> showRefineException <$> validate @(SymEqualTo "123") @Text Proxy "abc"
Just "RefineOtherException (SymEqualTo \"123\") \"abc does not equal the predicate\""

Since: 0.1.0.0

Char Predicates

This section models the boolean functions defined in Data.Char. See that module for more information regarding these definitions.

These instances are defined for Char and Word8, though they can be extended to String, Text, and ByteString via All.

Note: Although Char has instances for all of these predicates, some are missing for Word8 (and by extension ByteString). This is due to Word8's size restriction, i.e., a single byte 0 <= b <= 255. Thus, predicates that extend over the entirety of the unicode range (e.g. Alpha) do not have Word8 instances, as this could be misleading. For instance, consider the unicode character "ɦ" (U+0266):

>>> :{
  let txt = T.singleton ch -- ch == 0x0266 i.e. ɦ
   in isRight $ refine @(All Alpha) txt
:}
True

This Char is part of the alpha unicode category, so the refinement succeeds. On the other hand, suppose we have a Word8 instance that performs the obvious isAlpha . w2c:

>>> :{
  let bs = encodeUtf8 $ T.singleton ch
   in isRight $ refine @(All Alpha) bs
:}
False

The problem is that we are checking the underlying bytes if they satisfy isAlpha, but this is only true for Ascii alpha characters. Morally, our bytestring is this structure:

>>> BS.foldr (:) [] (encodeUtf8 $ T.singleton ch)
[201,166]

Our Alpha refinement fails because the individual byte components of "ɦ" are not themselves considered "alpha" characters (indeed this will only occur due to chance). These are the options:

Outlaw Word8/ByteString instances completely.

Provide the naive isAlpha . w2c implementation for Word8.

Implement ByteString instances by first converting to Text, i.e., do not use its underlying fold.

Provide Word8 instances only when they coincide with Char (i.e. ascii/latin1 predicates). In this case, ByteString works as expected; that is, we can make assertions based on the underlying bytes, but nothing that requires a specific encoding, and we do not get surprised by Text/ByteString mismatches.

Of these, only one and four are reasonable. Two is out because it can have confusing semantics (illustrated above).

Three is rejected because the API is no longer consistent, and we have to arbitrarily assume the ByteString shares its Text encoding (i.e. UTF-8).

One is defensible, but we choose option four, reasoning that it could be useful to assert that a given bytestring contains only ascii numbers or alpha characters while avoiding the pitfalls of reusing predicates intended for arbitrary unicode.

Unicode

These predicates are for unicode code points, hence they are available for Char (thus String, Text).

data Space

Predicate for a isSpace.

Examples

Expand

>>> validate @Space Proxy ' '
Nothing
>>> validate @Space Proxy '\r'
Nothing

>>> showRefineException <$> validate @Space Proxy 'a'
Just "RefineOtherException (Space) \"a is not a space character\""

Since: 0.1.0.0

data Lower

Predicate for isLower.

Examples

Expand

>>> validate @Lower Proxy 'c'
Nothing

>>> showRefineException <$> validate @Lower Proxy 'C'
Just "RefineOtherException (Lower) \"C is not lowercase\""

Since: 0.1.0.0

data Upper

Predicate for isUpper.

Examples

Expand

>>> validate @Upper Proxy 'C'
Nothing

>>> showRefineException <$> validate @Upper Proxy 'c'
Just "RefineOtherException (Upper) \"c is not uppercase\""

Since: 0.1.0.0

data Alpha

Predicate for isAlpha.

Examples

Expand

>>> validate @Alpha Proxy 'c'
Nothing

>>> validate @Alpha Proxy 'ɦ'
Nothing

>>> showRefineException <$> validate @Alpha Proxy '7'
Just "RefineOtherException (Alpha) \"7 is not an alphabetic character\""

Since: 0.1.0.0

data AlphaNum

Predicate for isAlphaNum.

Examples

Expand

>>> validate @AlphaNum Proxy 'a'
Nothing

>>> validate @AlphaNum Proxy '1'
Nothing

>>> showRefineException <$> validate @AlphaNum Proxy '!'
Just "RefineOtherException (AlphaNum) \"! is not an alpha-numeric character\""

Since: 0.1.0.0

data Letter

Predicate for isLetter.

Examples

Expand

>>> validate @Letter Proxy 'f'
Nothing

>>> showRefineException <$> validate @Letter Proxy '\r'
Just "RefineOtherException (Letter) \"\\r is not a letter\""

Since: 0.1.0.0

data Mark

Predicate for isMark.

Examples

Expand

>>> validate @Mark Proxy '\x20DD'
Nothing

>>> showRefineException <$> validate @Mark Proxy 'a'
Just "RefineOtherException (Mark) \"a is not a mark\""

Since: 0.1.0.0

data Number

Predicate for isNumber.

Examples

Expand

>>> validate @Number Proxy '2'
Nothing

>>> showRefineException <$> validate @Number Proxy 'a'
Just "RefineOtherException (Number) \"a is not a number\""

Since: 0.1.0.0

data Punctuation

Predicate for isPunctuation.

Examples

Expand

>>> validate @Punctuation Proxy '!'
Nothing

>>> showRefineException <$> validate @Punctuation Proxy 'a'
Just "RefineOtherException (Punctuation) \"a is not punctuation\""

Since: 0.1.0.0

data Symbol

Predicate for isSymbol.

Examples

Expand

>>> validate @Symbol Proxy '$'
Nothing

>>> showRefineException <$> validate @Symbol Proxy 'a'
Just "RefineOtherException (Symbol) \"a is not a symbol\""

Since: 0.1.0.0

data Separator

Predicate for isSeparator.

Examples

Expand

>>> validate @Separator Proxy ' '
Nothing

>>> showRefineException <$> validate @Separator Proxy 'a'
Just "RefineOtherException (Separator) \"a is not a separator\""

Since: 0.1.0.0

Ascii/Latin1

These predicates are for ascii/latin1. Thus they will work for Char (and String, Text) and Word8 (hence ByteString).

data Control

Predicate for isControl.

Examples

Expand

>>> validate @Control Proxy '\r'
Nothing

>>> showRefineException <$> validate @Control Proxy 'a'
Just "RefineOtherException (Control) \"a is not a control character\""

Since: 0.1.0.0

data Digit

Predicate for isDigit.

Examples

Expand

>>> validate @Digit Proxy '1'
Nothing

>>> showRefineException <$> validate @Digit Proxy 'a'
Just "RefineOtherException (Digit) \"a is not a digit\""

Since: 0.1.0.0

data OctDigit

Predicate for isOctDigit.

Examples

Expand

>>> validate @OctDigit Proxy '4'
Nothing

>>> showRefineException <$> validate @OctDigit Proxy '9'
Just "RefineOtherException (OctDigit) \"9 is not an octal digit\""

Since: 0.1.0.0

data HexDigit

Predicate for isHexDigit.

Examples

Expand

>>> validate @HexDigit Proxy '1'
Nothing

>>> validate @HexDigit Proxy 'f'
Nothing

>>> showRefineException <$> validate @HexDigit Proxy 'g'
Just "RefineOtherException (HexDigit) \"g is not a hexadecimal digit\""

Since: 0.1.0.0

data Ascii

Predicate for isAscii.

Examples

Expand

>>> validate @Ascii Proxy 'a'
Nothing

>>> showRefineException <$> validate @Ascii Proxy '\x20DD'
Just "RefineOtherException (Ascii) \"\\8413 is not ascii\""

Since: 0.1.0.0

data Latin1

Predicate for Latin1.

Examples

Expand

>>> validate @Latin1 Proxy 'a'
Nothing

>>> showRefineException <$> validate @Latin1 Proxy '\x20DD'
Just "RefineOtherException (Latin1) \"\\8413 is not latin1\""

Since: 0.1.0.0

data AsciiUpper

Predicate for isAsciiUpper.

Examples

Expand

>>> validate @AsciiUpper Proxy 'A'
Nothing

>>> showRefineException <$> validate @AsciiUpper Proxy 'a'
Just "RefineOtherException (AsciiUpper) \"a is not uppercase ascii\""

Since: 0.1.0.0

data AsciiLower

Predicate for isAsciiLower.

Examples

Expand

>>> validate @AsciiLower Proxy 'a'
Nothing

>>> showRefineException <$> validate @AsciiLower Proxy 'A'
Just "RefineOtherException (AsciiLower) \"A is not lowercase ascii\""

Since: 0.1.0.0

data AsciiAlpha

Predicate for isAscii and isAlpha, primarily intended for Word8. Redundant for Char, as this is equivalent to (Ascii && Alpha), but we include Char for completeness.

Examples

Expand

>>> validate @AsciiAlpha Proxy 'a'
Nothing

>>> showRefineException <$> validate @AsciiAlpha Proxy '1'
Just "RefineOtherException (AsciiAlpha) \"1 is not alpha ascii\""

>>> showRefineException <$> validate @AsciiAlpha Proxy 'ɦ'
Just "RefineOtherException (AsciiAlpha) \"\\614 is not alpha ascii\""

Since: 0.1.0.0

data AsciiAlphaNum

Predicate for isAscii and isAlphaNum, primarily intended for Word8. Redundant for Char, as this is equivalent to (Ascii && AlphaNum), but we include Char for completeness.

Examples

Expand

>>> validate @AsciiAlphaNum Proxy 'a'
Nothing

>>> validate @AsciiAlphaNum Proxy '1'
Nothing

>>> showRefineException <$> validate @AsciiAlphaNum Proxy '1'
Nothing

>>> showRefineException <$> validate @AsciiAlphaNum Proxy 'ɦ'
Just "RefineOtherException (AsciiAlphaNum) \"\\614 is not alpha-numeric ascii\""

Since: 0.1.0.0