Refined.Extras.Polymorphism

Description

Type families for defining functions that are polymorphic over predicates.

Since: 0.1.0.0

Documentation

The Implies type family lets us write functions that are polymorphic over predicates, i.e., we can define functions that require some predicate p, which can then be used by any refined term, as long as the term satisfies p . This means we can write APIs that are both safe and flexible.

type family Implies q p where ...

Implies q p raises a type error if q does not logically imply p. See ImpliesBool for when q implies p.

Examples

Expand

>>> :{
  let safeDiv :: (Implies p NonZero, Integral n) => n -> Refined p n -> n
      safeDiv x d = x `div` unrefine d
      --
      -- term does not have to exactly match Refined '[NonZero] n
      d :: Refined (NonZero && Positive) Int
      d = $$(refineTH 7)
   in safeDiv 14 d
:}
2

>>> :{
  -- total wrapper over Data.Text's partial splitOn.
  let safeSplitOn :: Implies p NonEmpty => Refined p Text -> Text -> [Text]
      safeSplitOn needle haystack = unrefine needle `splitOn` haystack
      --
      splitter = $$(refineTH @NonEmpty @Text ",")
   in safeSplitOn splitter "some,text,"
:}
["some","text",""]

Since: 0.1.0.0

Equations

Implies q p = ErrIfFalse (PredNotFound q p) (ImpliesBool q p)

type family ImpliesBool q p where ...

ImpliesBool q p returns 'True if q logically implies p, i.e., whenever q is true, p is also true.

NB. For a (possibly exclusive) disjunction to imply p, both clauses must individually imply p, since we cannot know which one is satisfied. That is,

\[ q_1 \vee q_2 \implies p \quad \iff \quad (q_1 \implies p) \wedge (q_2 \implies p) \]

When determining whether q implies p we are unfortunately limited by searching for syntactic, not sematic, equality. In particular,

  ((NonZero && NonNegative) && Even) /=> (NonZero && (NonNegative && Even))
  (NonZero || (NonNegative && Not NonNegative) || Even) /=> NonZero || Even

Thus it is advisable to write function constraints as simply as possible.

-- not this
foo :: (Implies p (NonZero && NonNegative) => Refined p a -> ...
-- prefer this
bar :: (Implies p NonZero, Implies p NonNegative) => Refined p a -> ...

Of course this transformation cannot be performed universally (e.g. A || B), and we do have

  (A && B) == (B && A), (A || B) == (B || A)

for convenience.

Examples

Expand

>>> :kind! ImpliesBool (Not NonZero) (NonNegative)
ImpliesBool (Not NonZero) (NonNegative) :: Bool
= 'False

>>> :kind! ImpliesBool (Not (Not NonNegative)) NonNegative
ImpliesBool (Not (Not NonNegative)) NonNegative :: Bool
= 'True

>>> :kind! ImpliesBool (NonZero || Alpha) NonZero
ImpliesBool (NonZero || Alpha) NonZero :: Bool
= 'False

>>> :kind! ImpliesBool (Alpha || (Alpha && Lower)) Alpha
ImpliesBool (Alpha || (Alpha && Lower)) Alpha :: Bool
= 'True

Since: 0.1.0.0

Equations

ImpliesBool p p = 'True

ImpliesBool q p = ImpliesCNF (ToCNF q) (ToCNF p)